// semana 19/Aula1 - Testes automatizados com Typescript/jest-template/src/models/User_2.ts
export enum LOCATION {
    EUA = "EUA",
    BRAZIL = "BRAZIL",
  }
  
export enum NACIONALITY {
    BRAZILIAN = "BRAZILIAN",
    AMERICAN = "AMERICAN",
  }
  
export interface User {
    name: string;
    age: number;
    nacionality: NACIONALITY;
  }
  
export interface Casino {
    name: string;
    location: LOCATION;
  }

export interface Result {
    brazilians: ResultItem;
    americans: ResultItem;
  }
  
export interface ResultItem {
    allowed: string[];
    unallowed: string[];
  }